requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]
python-jose
psycopg2-binary
pydantic
